+++ DB/ADMA.EWRS.Database/Weekly/Tables/SubjectStatuses.sql
@@ -1,7 +1,9 @@
 ﻿CREATE TABLE [Weekly].[SubjectStatuses] (
-    [SubjectStatus_Id] TINYINT      IDENTITY (1, 1) NOT NULL,
+    [SubjectStatus_Id] INT          IDENTITY (1, 1) NOT NULL,
     [Status]           VARCHAR (50) NOT NULL,
     [RowVersion]       ROWVERSION   NOT NULL,
     CONSTRAINT [PK_SubjectStatuses] PRIMARY KEY CLUSTERED ([SubjectStatus_Id] ASC)
 );
+
+
